requirements.txt
sentence-transformers
chromadb
pypdf
python-dotenv
groq